fixed Monitor UI
