+++ components/queueManager/Index.ts
@@ -7,5 +7,5 @@ export * from './NotManaged/NotManagedCalls'
 export * from './NotManaged/NotManagedCallsFilter'
 export * from './Realtime/RealTimeManagement'
 export * from './Summary/Summary'
-export * from './Monitor'
+export * from './Monitor/Monitor'
 export * from './QueueManagement/QueueManagementFilterOperators'
